Re-factor the psot index into a component.
