+++ src/components/posts/index/index.js
@@ -1,0 +1,15 @@
+/**
+ * External
+ */
+import Vue from 'vue';
+
+/**
+ * Internal
+ */
+import postIndex from './template.vue';
+
+/**
+ * Declare the component and export it
+ */
+Vue.component('post-index', postIndex);
+export default postIndex;
